Branch a packages (#3)
* feat: package progress

* Update .gitignore

* Update .gitignore

* Update .gitignore
