+++ data/lyra.txt
@@ -1,1 +1,3 @@
 T | 0 | final test
+T | 0 | Hello
+D | 0 | lol | tmr
