--- src/main/java/AddTodoCommand.java
@@ -1,15 +1,0 @@
-public class AddTodoCommand extends Command {
-    private final String description;
-    
-    public AddTodoCommand(String description) {
-        this.description = description;
-    }
-    
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws LyraException {
-        Task task = new Todo(description);
-        tasks.addTask(task);
-        storage.save(tasks.getTasks());
-        ui.showTaskAdded(task, tasks.getSize());
-    }
-}
