+++ src/main/java/duke/command/AddDeadlineCommand.java
@@ -1,0 +1,26 @@
+package duke.command;
+
+import duke.task.TaskList;
+import duke.task.Deadline;
+import duke.task.Task;
+import duke.util.Ui;
+import duke.util.Storage;
+import duke.exception.LyraException;
+
+public class AddDeadlineCommand extends Command {
+    private final String description;
+    private final String by;
+    
+    public AddDeadlineCommand(String description, String by) {
+        this.description = description;
+        this.by = by;
+    }
+    
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws LyraException {
+        Task task = new Deadline(description, by);
+        tasks.addTask(task);
+        storage.save(tasks.getTasks());
+        ui.showTaskAdded(task, tasks.getSize());
+    }
+}
