+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.task.TaskList;
+import duke.util.Ui;
+import duke.util.Storage;
+import duke.exception.LyraException;
+
 public abstract class Command {
     
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws LyraException;
