+++ src/main/java/duke/command/MarkCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.task.TaskList;
+import duke.util.Ui;
+import duke.util.Storage;
+import duke.exception.LyraException;
+
 public class MarkCommand extends Command {
     private final int taskIndex;
