+++ src/main/java/duke/exception/LyraException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class LyraException extends Exception {
     public LyraException(String message) {
         super(message);
