+++ src/main/java/duke/util/Parser.java
@@ -1,3 +1,16 @@
+package duke.util;
+
+import duke.command.Command;
+import duke.command.ExitCommand;
+import duke.command.ListCommand;
+import duke.command.AddTodoCommand;
+import duke.command.AddDeadlineCommand;
+import duke.command.AddEventCommand;
+import duke.command.MarkCommand;
+import duke.command.UnmarkCommand;
+import duke.command.DeleteCommand;
+import duke.exception.LyraException;
+
 public class Parser {
     
     public static Command parse(String fullCommand) throws LyraException {
